+++ Changelog.md
@@ -1,3 +1,12 @@
+# 2.0.0-dev15
+
+* `log` -> `info`
+* Add mitigation for KLayout None variables.
+* `logging` isolated from `common` into its own module
+* CLI now accepts either a value or a string for log levels.
+* CLI now prints help if no arguments are provided.
+* Fix issue where `rich` eats the cursor if it exits by interrupt.
+
 # 2.0.0-dev14
 
 * Multiple logging levels specified via CLI. Can also be set via `set_log_level` in the API.